Rename package and classes
