+++ shogun-lib/src/main/java/de/terrestris/shoguncore/graphql/fetcher/FileGraphQLDataFetcher.java
@@ -1,8 +1,8 @@
-package de.terrestris.shoguncore.graphql.fetchers;
+package de.terrestris.shoguncore.graphql.fetcher;
 
 import de.terrestris.shoguncore.model.File;
 import de.terrestris.shoguncore.service.FileService;
 import org.springframework.stereotype.Component;
 
 @Component
-public class FileGraphQLProvider extends BaseGraphQLDataFetcher<File, FileService> { }
+public class FileGraphQLDataFetcher extends BaseGraphQLDataFetcher<File, FileService> { }
