+++ shogun-lib/src/main/java/de/terrestris/shoguncore/graphql/fetcher/ImageFileGraphQLDataFetcher.java
@@ -1,8 +1,8 @@
-package de.terrestris.shoguncore.graphql.fetchers;
+package de.terrestris.shoguncore.graphql.fetcher;
 
 import de.terrestris.shoguncore.model.ImageFile;
 import de.terrestris.shoguncore.service.ImageFileService;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ImageFileGraphQLProvider extends BaseGraphQLDataFetcher<ImageFile, ImageFileService> { }
+public class ImageFileGraphQLDataFetcher extends BaseGraphQLDataFetcher<ImageFile, ImageFileService> { }
