+++ shogun-lib/src/main/java/de/terrestris/shoguncore/graphql/fetcher/LayerGraphQLDataFetcher.java
@@ -1,8 +1,8 @@
-package de.terrestris.shoguncore.graphql.fetchers;
+package de.terrestris.shoguncore.graphql.fetcher;
 
 import de.terrestris.shoguncore.model.Layer;
 import de.terrestris.shoguncore.service.LayerService;
 import org.springframework.stereotype.Component;
 
 @Component
-public class LayerGraphQLProvider extends BaseGraphQLDataFetcher<Layer, LayerService> { }
+public class LayerGraphQLDataFetcher extends BaseGraphQLDataFetcher<Layer, LayerService> { }
