+++ shogun-lib/src/main/java/de/terrestris/shoguncore/graphql/fetcher/RoleGraphQLDataFetcher.java
@@ -1,8 +1,8 @@
-package de.terrestris.shoguncore.graphql.fetchers;
+package de.terrestris.shoguncore.graphql.fetcher;
 
 import de.terrestris.shoguncore.model.Role;
 import de.terrestris.shoguncore.service.RoleService;
 import org.springframework.stereotype.Component;
 
 @Component
-public class RoleGraphQLProvider extends BaseGraphQLDataFetcher<Role, RoleService> { }
+public class RoleGraphQLDataFetcher extends BaseGraphQLDataFetcher<Role, RoleService> { }
