+++ shogun-lib/src/main/java/de/terrestris/shoguncore/graphql/fetcher/security/IdentityGraphQLProvider.java
@@ -1,6 +1,6 @@
-package de.terrestris.shoguncore.graphql.fetchers.security;
+package de.terrestris.shoguncore.graphql.fetcher.security;
 
-import de.terrestris.shoguncore.graphql.fetchers.BaseGraphQLDataFetcher;
+import de.terrestris.shoguncore.graphql.fetcher.BaseGraphQLDataFetcher;
 import de.terrestris.shoguncore.model.security.Identity;
 import de.terrestris.shoguncore.service.security.IdentityService;
 import org.springframework.stereotype.Component;
